refactor: add new file with the whole story, and one file is clean
